+++ src/lib.rs
@@ -1,4 +1,7 @@
+pub mod common;
+pub mod compute;
 pub mod frost;
 pub mod schnorr;
+pub mod traits;
 pub mod util;
 pub mod vss;
