+++ src/traits.rs
@@ -1,0 +1,6 @@
+pub trait Signer {
+    fn new(num_keys: usize) -> Self;
+    fn load(path: &str) -> Self;
+
+    fn save(path: &str);
+}
